+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Secondary device attributes escape (`CSI > 0 c`)
 
+### Changed
+
+- Cursors are now inverted when their fixed color is similar to the cell's background
+
 ## 0.5.0-dev
 
 ### Packaging